Updated scripts to match the structure in deploy
