--- src/scripts/preload.js
@@ -1,2 +1,0 @@
-//TODO (may call the file system functionality)
-console.log("IM PRELOADED");
